This commit was generated by cvs2svn to compensate for changes in r5879, which included commits to RCS files with non-trunk default branches.
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@5880 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/arch/win32/mouse.c
@@ -165,7 +165,7 @@ HRESULT result;
 
     result=IDirectInputDevice_SetCooperativeLevel(di_mouse,ui_active_window,DISCL_EXCLUSIVE|DISCL_FOREGROUND);
     if (result!=DI_OK) {
-        log_debug("Warning: couldn't set cooperation level of mice to exclusive! %d",ui_active_window);
+        log_debug("Warning: couldn't set cooperation level of mice to exclusive! %i",(int)ui_active_window);
         di_mouse=NULL;
     }
 }
